Update telegram_bot
This code is created as an example of how the bot works with a database. 
First, we load the necessary libraries for the bot to work. For security purposes, the keys to the database and the bot token are specified in the .env file so that they cannot be accessed. To work with the .env file we use the sqlite3 library. Next, we create the necessary tables in the database. Then we prescribe the necessary commands and conditions for the bot. When specifying a text the bot saves the user's chat id and the text of the message, when uploading a photo it saves it to a folder and makes a record in the database with the video as well.